+++ src/gitstore/__init__.py
@@ -1,13 +1,13 @@
 from .repo import GitStore, ReflogEntry
 from .mirror import MirrorDiff, RefChange
-from .fs import FS, retry_write
+from .fs import FS, WriteEntry, retry_write
 from .tree import WalkEntry
 from .exceptions import StaleSnapshotError
 from .copy import ChangeReport, ChangeAction, ChangeError, FileEntry, FileType, disk_glob
 
 __all__ = [
     "GitStore", "ReflogEntry", "MirrorDiff", "RefChange",
-    "FS", "retry_write", "StaleSnapshotError",
+    "FS", "WriteEntry", "retry_write", "StaleSnapshotError",
     "ChangeReport", "ChangeAction", "ChangeError", "FileEntry", "FileType", "disk_glob",
     "WalkEntry",
 ]